+++ CHANGELOG.md
@@ -10,14 +10,18 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 <!-- TODO: change `@added unknown` tag due to the actual version -->
 <!--       during new scriptlets or redirects releasing -->
 
-## [Unreleased]
+## [v2.3.1] - 2026-03-24
+
+### Changed
+
+- Updated [@adguard/agtree] to `4.0.3`.
 
 ### Fixed
 
 - `disable-newtab-links` not preventing clicks when the element has its own `click` handler
   added via `addEventListener` [#483].
 
-[Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.3.0...HEAD
+[v2.3.1]: https://github.com/AdguardTeam/Scriptlets/compare/v2.3.0...v2.3.1
 [#483]: https://github.com/AdguardTeam/Scriptlets/issues/483
 
 ## [v2.3.0] - 2026-03-18